# openapi/wallet-api.yaml
openapi: 3.1.0
info:
  title: Universal AI Wallet — Wallet API
  version: 1.0.0
  description: Account, balance, keys, and top-ups.
servers:
  - url: https://api.example.com/wallet/v1
paths:
  /auth/register:
    post:
      operationId: register
      summary: Create account
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      operationId: login
      summary: Login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /me:
    get:
      operationId: getMe
      summary: Current user profile
      tags: [User]
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /wallet:
    get:
      operationId: getWallet
      summary: Wallet balance
      tags: [Wallet]
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
...
  /auth/oauth/google:
    get:
      operationId: googleLoginRedirect
      summary: Redirect to Google OAuth consent
      tags: [Auth]
      security: []
      responses:
        '302':
          description: Redirect to Google
        '503':
          description: Google OAuth not configured
    post:
      operationId: googleLoginCallback
      summary: Exchange Google code for a wallet session
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, state]
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        '200':
          description: Wallet session (same shape as email/password login)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid state or exchange failure
  /apps:
    get:
      operationId: listConnectedApps
      summary: List the user's connected apps
      tags: [ConnectedApps]
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Connected apps
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConnectedApp'
  /apps/{clientId}/connect:
    post:
      operationId: connectApp
      summary: Connect / set a per-app spend cap
      tags: [ConnectedApps]
      security:
        - JwtAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spend_limit_microdollars:
                  type: integer
                  minimum: 0
                  nullable: true
                reset_period:
                  type: string
                  enum: [monthly, lifetime]
                  default: monthly
                display_name:
                  type: string
      responses:
        '201':
          description: Connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedApp'
  /apps/{installId}:
    get:
      operationId: getConnectedApp
      summary: Connected app detail
      tags: [ConnectedApps]
      security:
        - JwtAuth: []
      parameters:
        - name: installId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedApp'
    patch:
      operationId: updateAllowance
      summary: Update the per-app spend cap
      tags: [ConnectedApps]
      security:
        - JwtAuth: []
      parameters:
        - name: installId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spend_limit_microdollars:
                  type: integer
                  minimum: 0
                  nullable: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedApp'
    delete:
      operationId: revokeApp
      summary: Revoke app access and all refresh tokens
      tags: [ConnectedApps]
      security:
        - JwtAuth: []
      parameters:
        - name: installId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Revoked
  /partner/{partnerSlug}/apps:
    post:
      operationId: createAppRegistration
      summary: Register an OAuth client (partner admin)
      tags: [PartnerApps]
      security:
        - PartnerAdmin: []
      parameters:
        - name: partnerSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, redirect_uris]
              properties:
                name:
                  type: string
                redirect_uris:
                  type: array
                  items:
                    type: string
                  minItems: 1
                scopes:
                  type: array
                  items:
                    type: string
                logo_url:
                  type: string
      responses:
        '201':
          description: Created — client_secret returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppRegistrationCreated'
    get:
      operationId: listAppRegistrations
      summary: List app registrations (partner admin)
      tags: [PartnerApps]
      security:
        - PartnerAdmin: []
      parameters:
        - name: partnerSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AppRegistration'

  # --- Root-mounted OAuth2/OIDC endpoints (no /wallet/v1 prefix) ---
...
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuthBearer:
      type: http
      scheme: bearer
      description: OAuth2 app access token (JWT with app_install_id claim)
    PartnerAdmin:
      type: apiKey
      in: header
      name: X-Partner-Admin-Token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    Wallet:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        balance_microdollars:
          type: integer
        held_microdollars:
          type: integer
        available_microdollars:
          type: integer
        currency:
          type: string
          enum: [USD]
        low_balance_threshold_microdollars:
          type: integer
...
    ConnectedApp:
      type: object
      properties:
        install_id:
          type: string
          format: uuid
        client_id:
          type: string
        app_name:
          type: string
        display_name:
          type: string
          nullable: true
        spend_limit_microdollars:
          type: integer
          nullable: true
        allowance_spent_microdollars:
          type: integer
        allowance_reset_period:
          type: string
          enum: [monthly, lifetime]
        consented_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
    AppRegistration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        partner_account_id:
          type: string
          format: uuid
        name:
          type: string
        client_id:
          type: string
        redirect_uris:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        logo_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AppRegistrationCreated:
      allOf:
        - $ref: '#/components/schemas/AppRegistration'
        - type: object
          required: [client_secret]
          properties:
            client_secret:
              type: string
              description: Full secret — only returned at creation/rotation
...